fix: rename md to mdx
