+++ packages/forma-36-react-components/src/components/Button/README.mdx
@@ -65,3 +65,5 @@ Button component is used to trigger an action or event. By default button is imp
 When a button has focus, both the Space and Enter keys will activate the button. If the button is disabled, then the disabled attribute is provided.
 Contentful buttons are checked in terms of color contrast and pass all the requirements.
 Buttons have a focus state when using keyboard navigation.
+
+export default ({ children }) => <>{children}</>
