+++ packages/forma-36-react-components/src/components/DateTime/DateTime/DateTime.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import DateTime, { DateTimeProps } from './DateTime';
 
-import notes from '../README.md';
+import notes from '../README.mdx';
 
 const exampleDate = '2020-08-13T13:45:56.0123Z';
