+++ packages/forma-36-react-components/src/components/DateTime/RelativeDate/RelativeDate.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import RelativeDate, { RelativeDateProps } from './RelativeDate';
 
-import notes from '../README.md';
+import notes from '../README.mdx';
 
 const THREE_MINUTES = 3 * 60 * 1000;
