+++ packages/forma-36-react-components/src/components/Pill/README.mdx
