+++ packages/forma-36-react-components/src/components/ProductIcon/README.mdx
